// frontend/src/app/core/api/model/shareLinkResponse.ts
/**
 * Rev-PasswordManager API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ShareLinkResponse { 
    shareId?: number;
    shareToken?: string;
    shareUrl?: string;
    encryptionKey?: string;
    vaultEntryTitle?: string;
    recipientEmail?: string;
    permission?: string;
    maxViews?: number;
    viewCount?: number;
    expiresAt?: string;
    createdAt?: string;
    revoked?: boolean;
}